+++ lib/mutant.rb
@@ -48,6 +48,9 @@ require 'mutant/ast/nodes'
 require 'mutant/ast/named_children'
 require 'mutant/ast/node_predicates'
 require 'mutant/ast/meta'
+require 'mutant/ast/meta/send'
+require 'mutant/ast/meta/optarg'
+require 'mutant/ast/meta/resbody'
 require 'mutant/actor'
 require 'mutant/actor/receiver'
 require 'mutant/actor/sender'
